requirements.txt
# Flask==1.0.2
# Flask-WTF==0.14.2
# WTForms==2.2.1
# google-cloud-datastore==1.6.0

asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
configparser==3.5.0
crcmod==1.7
cryptography==2.1.4
enum34==1.1.6
Flask==1.0.2
Flask-WTF==0.14.2
futures==3.2.0
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
ipaddress==1.0.17
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycrypto==2.6.1
pygobject==3.26.1
pytz==2018.5
pyxdg==0.25
requests==2.19.1
rsa==4.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.23
vboxapi==1.0
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1